feat(stylesheets): use separate meta build since its internal for now | hj | #4523-follow
