+++ .gitignore
@@ -114,6 +114,7 @@ packages/styles/projects/scss/meta/
 packages/styles/projects/tailwindcss/meta/
 packages/styles/projects/vanilla-extract/meta/
 packages/styles/projects/emotion/meta/
+packages/components/projects/stylesheets/meta/
 
 # Agent Skills, ignored for now for testing and review of the skills. Remove once we decided on what skills to keep.
 .claude/
