+++ packages/components/projects/stylesheets/package.json
@@ -9,29 +9,31 @@
   "sideEffects": false,
   "exports": {
     ".": {
-      "types": "./dist/esm/index.d.ts",
-      "import": "./dist/esm/index.mjs",
-      "require": "./dist/cjs/index.cjs"
+      "style": "./dist/index.css",
+      "types": "./meta/esm/index.d.ts",
+      "import": "./meta/esm/index.mjs",
+      "require": "./meta/cjs/index.cjs"
     },
     "./meta": {
-      "types": "./dist/esm/meta.d.ts",
-      "import": "./dist/esm/meta.mjs",
-      "require": "./dist/cjs/meta.cjs"
+      "types": "./meta/esm/meta.d.ts",
+      "import": "./meta/esm/meta.mjs",
+      "require": "./meta/cjs/meta.cjs"
     },
     "./*": "./dist/*"
   },
-  "main": "./dist/cjs/index.cjs",
-  "module": "./dist/esm/index.mjs",
-  "types": "./dist/esm/index.d.ts",
+  "main": "./meta/cjs/index.cjs",
+  "module": "./meta/esm/index.mjs",
+  "types": "./meta/esm/index.d.ts",
   "style": "./dist/index.css",
   "files": [
-    "./dist/"
+    "./dist/",
+    "./meta/"
   ],
   "scripts": {
-    "build": "npm run clean:build && npm run build:constants && npm run build:package && npm run build:css && npm run build:font-face && npm run build:index && npm run build:skill",
-    "build-prod": "npm run clean:build && npm run build:constants && npm run build:package && npm run build:css && npm run build-prod:font-face && npm run build:index && npm run build:skill",
+    "build": "npm run clean:build && npm run build:constants && npm run build:meta && npm run build:css && npm run build:font-face && npm run build:index && npm run build:skill",
+    "build-prod": "npm run clean:build && npm run build:constants && npm run build:meta && npm run build:css && npm run build-prod:font-face && npm run build:index && npm run build:skill",
     "build:constants": "tsx scripts/buildCssVariableConstants.ts",
-    "build:package": "rollup -c rollup.config.mjs",
+    "build:meta": "rollup -c rollup.config.meta.mjs",
     "build:css": "tsx scripts/buildStylesheetsCss.ts",
     "build:skill": "tsx scripts/build-skill.ts",
     "build:font-face": "NODE_ENV=development tsx scripts/buildFontFaceCss.ts",
@@ -41,7 +43,7 @@
     "build:index:cdn": "mkdir -p dist/cn && cat dist/normalize.css dist/variables.css dist/color-scheme.css dist/cn/font-face.css > dist/cn/index.css",
     "test": "npm run test:unit",
     "test:unit": "vitest --config tests/unit/config/vitest.config.ts",
-    "clean:build": "rm -rf ./dist"
+    "clean:build": "rm -rf ./dist ./meta"
   },
   "dependencies": {
     "@porsche-design-system/font-face": "0.0.0",
